style(backend): refactor fn create_instance
