+++ src-tauri/src/instance/commands.rs
@@ -1,25 +1,20 @@
 use super::helpers::loader::fabric::remove_fabric_api_mods;
 use crate::error::SJMCLResult;
 use crate::instance::constants::TRANSLATION_CACHE_EXPIRY_HOURS;
-use crate::instance::helpers::client_json::{replace_native_libraries, McClientInfo};
+use crate::instance::helpers::client_json::McClientInfo;
 use crate::instance::helpers::game_version::{build_game_version_cmp_fn, compare_game_versions};
 use crate::instance::helpers::loader::common::{execute_processors, install_mod_loader};
 use crate::instance::helpers::loader::forge::InstallProfile;
-use crate::instance::helpers::loader::optifine::{
-  download_optifine_installer, finish_optifine_install,
-};
+use crate::instance::helpers::loader::optifine::finish_optifine_install;
 use crate::instance::helpers::misc::{
   get_instance_game_config, get_instance_subdir_path_by_id, get_instance_subdir_paths,
   refresh_and_update_instances, unify_instance_name,
 };
-use crate::instance::helpers::modpack::misc::{
-  extract_overrides, get_download_params, ModpackMetaInfo,
-};
+use crate::instance::helpers::modpack::misc::ModpackMetaInfo;
 use crate::instance::helpers::mods::common::{
   add_local_mod_translations, compress_icon, get_mod_info_from_dir, get_mod_info_from_jar,
   LocalModTranslationEntry, LocalModTranslationsCache,
 };
-use crate::instance::helpers::options_txt::get_zh_hans_lang_tag;
 use crate::instance::helpers::resourcepack::{
   load_resourcepack_from_dir, load_resourcepack_from_zip,
 };
@@ -29,12 +24,10 @@ use crate::instance::helpers::server::{
 use crate::instance::helpers::world::{load_level_data_from_nbt, load_world_info_from_dir};
 use crate::instance::models::misc::{
   Instance, InstanceError, InstanceSubdirType, InstanceSummary, LocalModInfo, ModLoader,
-  ModLoaderStatus, ModLoaderType, OptiFine, ResourcePackInfo, SchematicInfo, ScreenshotInfo,
-  ShaderPackInfo,
+  ModLoaderStatus, ModLoaderType, ResourcePackInfo, SchematicInfo, ScreenshotInfo, ShaderPackInfo,
 };
 use crate::instance::models::world::base::WorldInfo;
 use crate::instance::models::world::level::LevelData;
-use crate::launch::helpers::file_validator::{get_invalid_assets, get_invalid_library_files};
 use crate::launcher_config::helpers::misc::get_global_game_config;
 use crate::launcher_config::models::{GameConfig, GameDirectory, LauncherConfig};
 use crate::partial::{PartialError, PartialUpdate};
@@ -44,7 +37,6 @@ use crate::resource::models::{
 };
 use crate::storage::{load_json_async, save_json_async, Storage};
 use crate::tasks::commands::schedule_progressive_task_group;
-use crate::tasks::download::DownloadParam;
 use crate::tasks::PTaskParam;
 use crate::utils::fs::{
   copy_whole_dir, create_url_shortcut, generate_unique_filename, get_files_with_regex,
@@ -60,10 +52,8 @@ use std::sync::{Arc, Mutex};
 use std::time::SystemTime;
 use tauri::State;
 use tauri::{AppHandle, Manager};
-use tauri_plugin_http::reqwest;
 use tokio;
 use tokio::sync::Semaphore;
-use url::Url;
 use zip::read::ZipArchive;
 
 #[tauri::command]
@@ -872,190 +862,29 @@ pub async fn create_instance(
   modpack_path: Option<String>,
   is_install_fabric_api: Option<bool>,
 ) -> SJMCLResult<()> {
-  let client = app.state::<reqwest::Client>();
-  let launcher_config_state = app.state::<Mutex<LauncherConfig>>();
-  // Get priority list
-  let priority_list = {
-    let launcher_config = launcher_config_state.lock()?;
-    get_source_priority_list(&launcher_config)
-  };
-
-  // Ensure the instance name is unique
-  let version_path = directory.dir.join("versions").join(&name);
-  if version_path.exists() {
-    return Err(InstanceError::ConflictNameError.into());
-  }
-  let optifine_info = optifine.as_ref().map(|info| OptiFine {
-    filename: info.filename.clone(),
-    version: format!("{}_{}", info.r#type, info.patch),
-    status: ModLoaderStatus::NotDownloaded,
-  });
-
-  // Create instance config
-  let instance = Instance {
-    id: format!("{}:{}", directory.name, name.clone()),
-    name: name.clone(),
-    version: game.id.clone(),
-    version_path: version_path.clone(),
-    mod_loader: ModLoader {
-      loader_type: mod_loader.loader_type.clone(),
-      status: if matches!(
-        mod_loader.loader_type,
-        ModLoaderType::Unknown | ModLoaderType::Fabric
-      ) {
-        ModLoaderStatus::Installed
-      } else {
-        ModLoaderStatus::NotDownloaded
-      },
-      version: mod_loader.version.clone(),
-      branch: mod_loader.branch.clone(),
-    },
-    optifine: optifine_info,
+  let path = directory.dir.join("versions").join(&name);
+  match crate::instance::helpers::instance_creator::create_instance(
+    app,
+    directory,
+    name,
     description,
     icon_src,
-    starred: false,
-    play_time: 0,
-    use_spec_game_config: false,
-    spec_game_config: None,
-  };
-
-  // Download version info
-  let mut version_info = client
-    .get(&game.url)
-    .send()
-    .await
-    .map_err(|_| InstanceError::NetworkError)?
-    .json::<McClientInfo>()
-    .await
-    .map_err(|_| InstanceError::ClientJsonParseError)?;
-
-  version_info.id = name.clone();
-  version_info.jar = Some(name.clone());
-
-  // convert vanilla version info to vanilla patch
-  let mut vanilla_patch = version_info.clone();
-  vanilla_patch.id = "game".to_string();
-  vanilla_patch.version = Some(game.id.clone());
-  vanilla_patch.inherits_from = None;
-  vanilla_patch.priority = Some(0);
-  version_info.patches.push(vanilla_patch);
-
-  let mut task_params = Vec::<PTaskParam>::new();
-
-  // Download client (use task)
-  let client_download_info = version_info
-    .downloads
-    .get("client")
-    .ok_or(InstanceError::ClientJsonParseError)?;
-
-  task_params.push(PTaskParam::Download(DownloadParam {
-    src: Url::parse(&client_download_info.url.clone())
-      .map_err(|_| InstanceError::ClientJsonParseError)?,
-    dest: instance.version_path.join(format!("{}.jar", name)),
-    filename: None,
-    sha1: Some(client_download_info.sha1.clone()),
-  }));
-  let subdirs = get_instance_subdir_paths(
-    &app,
-    &instance,
-    &[
-      &InstanceSubdirType::Libraries,
-      &InstanceSubdirType::Assets,
-      &InstanceSubdirType::Mods,
-    ],
+    game,
+    mod_loader,
+    optifine,
+    modpack_path,
+    is_install_fabric_api,
   )
-  .ok_or(InstanceError::InstanceNotFoundByID)?;
-  let [libraries_dir, assets_dir, mods_dir] = subdirs.as_slice() else {
-    return Err(InstanceError::InstanceNotFoundByID.into());
-  };
-
-  replace_native_libraries(&app, &mut version_info, &instance)
-    .await
-    .map_err(|_| InstanceError::ClientJsonParseError)?;
-
-  // We only download libraries if they are invalid (not already downloaded)
-  task_params.extend(
-    get_invalid_library_files(priority_list[0], libraries_dir, &version_info, false).await?,
-  );
-
-  // We only download assets if they are invalid (not already downloaded)
-  task_params
-    .extend(get_invalid_assets(&app, &version_info, priority_list[0], assets_dir, false).await?);
-
-  // download loader (installer)
-  if instance.mod_loader.loader_type != ModLoaderType::Unknown {
-    install_mod_loader(
-      app.clone(),
-      &priority_list,
-      &instance.version,
-      &instance.mod_loader,
-      libraries_dir.to_path_buf(),
-      mods_dir.to_path_buf(),
-      &mut version_info,
-      &mut task_params,
-      is_install_fabric_api,
-    )
-    .await?;
-  }
-
-  if let Some(info) = optifine.as_ref() {
-    download_optifine_installer(
-      &instance.version,
-      info,
-      libraries_dir.to_path_buf(),
-      &mut task_params,
-    )
-    .await?;
-  }
-
-  // If modpack path is provided, install it
-  if let Some(modpack_path) = modpack_path {
-    let path = PathBuf::from(modpack_path);
-    let file = fs::File::open(&path).map_err(|_| InstanceError::FileNotFoundError)?;
-    task_params.extend(get_download_params(&app, &file, &version_path).await?);
-    extract_overrides(&file, &version_path)?;
-  }
-
-  schedule_progressive_task_group(
-    app.clone(),
-    format!("game-client?{}", name),
-    task_params,
-    true,
-  )
-  .await?;
-
-  // Optionally skip first-screen options by adding options.txt (available for zh-Hans only)
-  let (language, skip_first_screen_options) = {
-    let launcher_config = launcher_config_state.lock()?;
-    (
-      launcher_config.general.general.language.clone(),
-      launcher_config
-        .general
-        .functionality
-        .skip_first_screen_options,
-    )
-  };
-  if language == "zh-Hans" && skip_first_screen_options {
-    if let Some(lang_code) = get_zh_hans_lang_tag(&instance.version, &app).await {
-      let options_path = get_instance_subdir_paths(&app, &instance, &[&InstanceSubdirType::Root])
-        .ok_or(InstanceError::InstanceNotFoundByID)?[0]
-        .join("options.txt");
-      if !options_path.exists() {
-        fs::write(options_path, format!("lang:{}\n", lang_code))
-          .map_err(|_| InstanceError::FileCreationFailed)?;
+  .await
+  {
+    Ok(_) => Ok(()),
+    Err(e) => {
+      if e != InstanceError::ConflictNameError.into() && path.exists() {
+        let _ = fs::remove_dir_all(&path);
       }
+      Err(e.into())
     }
   }
-
-  // Save the edited client json
-  save_json_async(&version_info, &version_path.join(format!("{}.json", name))).await?;
-  // Save the SJMCL instance config json
-  instance
-    .save_json_cfg()
-    .await
-    .map_err(|_| InstanceError::FileCreationFailed)?;
-
-  Ok(())
 }
 
 #[tauri::command]
