+++ src-tauri/src/instance/helpers/mod.rs
@@ -2,6 +2,7 @@ pub mod asset_index;
 pub mod client_jar;
 pub mod client_json;
 pub mod game_version;
+pub mod instance_creator;
 pub mod loader;
 pub mod misc;
 pub mod modpack;
